Confirmar cerrar sesion

+++ app/src/main/java/com/ivione93/hojames/ui/profile/ProfileFragment.java
@@ -1,12 +1,13 @@
 package com.ivione93.hojames.ui.profile;
 
 import android.app.ProgressDialog;
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -158,15 +159,24 @@ public class ProfileFragment extends Fragment {
         }
         if (item.getItemId() == R.id.menu_log_out) {
             // Borrado datos inicio de sesion
-            SharedPreferences.Editor prefs = getActivity().getSharedPreferences(getString(R.string.prefs_file), Context.MODE_PRIVATE).edit();
-            prefs.clear();
-            prefs.apply();
+            new AlertDialog.Builder(getActivity())
+                    .setTitle("Salir")
+                    .setMessage("¿Está seguro que quiere cerrar sesión?")
+                    .setCancelable(false)
+                    .setPositiveButton("Salir", (dialogInterface, i) -> {
+                        SharedPreferences.Editor prefs = getActivity().getSharedPreferences(getString(R.string.prefs_file), Context.MODE_PRIVATE).edit();
+                        prefs.clear();
+                        prefs.apply();
 
-            FirebaseAuth.getInstance().signOut();
-            mGoogleSignInClient.signOut();
+                        FirebaseAuth.getInstance().signOut();
+                        mGoogleSignInClient.signOut();
 
-            Intent mainIntent = new Intent(getActivity().getApplicationContext(), AuthActivity.class);
-            startActivity(mainIntent);
+                        Intent mainIntent = new Intent(getActivity().getApplicationContext(), AuthActivity.class);
+                        startActivity(mainIntent);
+                        getActivity().finish();
+                    })
+                    .setNegativeButton("Cancelar", (dialogInterface, i) -> dialogInterface.dismiss())
+                    .show();
         }
         return super.onOptionsItemSelected(item);
     }
